chore(30574): reorganise the combiner files
