+++ hivemq-edge/src/frontend/src/modules/Mappings/combiner/CombinedEntitySelect.tsx
