+++ hivemq-edge/src/frontend/src/modules/Mappings/combiner/DataCombiningTableField.spec.cy.tsx
