+++ hivemq-edge/src/frontend/src/modules/Mappings/combiner/DestinationSchemaLoader.tsx
