+++ hivemq-edge/src/frontend/src/modules/Mappings/combiner/EntityReferenceTableWidget.spec.cy.tsx
