+++ hivemq-edge/src/frontend/src/modules/Mappings/combiner/EntityReferenceTableWidget.tsx
