+++ hivemq-edge/src/frontend/src/modules/Mappings/combiner/EntityRenderer.tsx
