+++ hivemq-edge/src/frontend/src/modules/Mappings/combiner/PrimarySelect.tsx
